% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsymmetricmatrix.R
\name{rsymmetricmatrix}
\alias{rsymmetricmatrix}
\title{Quickly Generate a Symmetric Matrix for Testing and Examples}
\usage{
rsymmetricmatrix(p, min = 0, max = 1)
}
\arguments{
\item{p}{The desired dimension of the matrix}

\item{min}{The minimum of the uniform distribution.}

\item{max}{The maximum of the uniform distribution}
}
\value{
A \code{p} x \code{p} symmetric matrix.
}
\description{
A simple function for generating a symmetric matrix for use in examples.
The diagonal, and upper-triangular elements of the matrix are simulated independently from a uniform distribution. The lower-triangle of the output matrix is copied from the upper-triangle.
These matrices \strong{do not} represent the full range of possible symmetric matrices.
}
\examples{
rsymmetricmatrix(5)
}
